backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Async support
aiofiles>=23.2.1

# Redis and Celery
redis>=5.0.1
celery>=5.3.6

# Image processing (for thumbnails, not OCR)
pillow>=10.2.0

# PDF processing
pdfplumber>=0.11.0

# Fuzzy matching
rapidfuzz>=3.6.1

# HTTP client (for MinerU, Ollama, Open Food Facts)
httpx>=0.26.0

# Configuration
python-dotenv>=1.0.1
pydantic-settings>=2.1.0

# Logging
structlog>=24.1.0

# Date/time utilities
python-dateutil>=2.8.2

# Testing (dev dependencies)
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0
aiosqlite>=0.22.0

# Code quality (dev dependencies)
ruff>=0.2.0
mypy>=1.8.0
